Show errors on the project in lsp-treemacs-error-list
At least JDTLS also reports errors directly on the project (rather than just files and folders under the project). These were previously not shown at all in the lsp-treemacs error list, leading to weird cases where there was an entry with an error count but no children. This changes it so the errors on the project are shown directly.